VDB-437 - Update ilk query (#33)
* Remove ilk query dependence on pit contract
* Remove ilk query dependence on take and ink fields
* Add ilk dust to ilk query

+++ db/migrations/20190308111729_create_ilk_history_query.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 -- +goose StatementBegin
-create or replace function maker.get_ilk_history_before_block(block_number numeric, ilk_id int)
+create or replace function maker.get_ilk_history_before_block(block_number bigint, ilk_id int)
   returns setof maker.ilk_state as $$
 DECLARE
   r record;
@@ -8,7 +8,7 @@ BEGIN
   FOR r IN SELECT * FROM maker.get_ilk_blocks_before($1, $2)
   LOOP
     RETURN QUERY
-    SELECT * FROM maker.get_ilk_at_block_number(r.block_number::numeric, $2::integer);
+    SELECT * FROM maker.get_ilk_at_block_number(r.block_number::bigint, $2::integer);
   END LOOP;
 END;
 $$
@@ -16,4 +16,4 @@ LANGUAGE plpgsql;
 -- +goose StatementEnd
 
 -- +goose Down
-drop function  maker.get_ilk_history_before_block(block_number numeric, ilk_id int);
+drop function  maker.get_ilk_history_before_block(block_number bigint, ilk_id int);
